+++ src/database/list/create-reference.ts
@@ -1,5 +1,5 @@
 import { DatabaseQuery, AngularFireList, ChildEvent } from '../interfaces';
-import { createLoadedChanges, loadedSnapshotChanges } from './loaded';
+import { snapshotChanges } from './snapshot-changes';
 import { createStateChanges } from './state-changes';
 import { createAuditTrail } from './audit-trail';
 import { createDataOperationMethod } from './data-operation';
@@ -12,11 +12,13 @@ export function createListReference<T>(query: DatabaseQuery): AngularFireList<T>
     set: createDataOperationMethod<T>(query.ref, 'set'),
     push: (data: T) => query.ref.push(data),
     remove: createRemoveMethod(query.ref),
-    snapshotChanges: createLoadedChanges(query),
+    snapshotChanges(events?: ChildEvent[]) {
+      return snapshotChanges(query, events);
+    },
     stateChanges: createStateChanges(query),
     auditTrail: createAuditTrail(query),
     valueChanges<T>(events?: ChildEvent[]) { 
-      return loadedSnapshotChanges(query, events)
+      return snapshotChanges(query, events)
         .map(actions => actions.map(a => a.payload!.val())); 
     }
   }
