+++ src/root.spec.js
@@ -8,7 +8,6 @@ export * from './packages-dist/database/database.spec';
 export * from './packages-dist/database/utils.spec';
 export * from './packages-dist/database/observable/fromRef.spec';
 export * from './packages-dist/database/list/changes.spec';
-export * from './packages-dist/database/list/loaded.spec';
 export * from './packages-dist/database/list/snapshot-changes.spec';
 export * from './packages-dist/database/list/state-changes.spec';
 export * from './packages-dist/database/list/audit-trail.spec';
